# Send errors from filters to the application HttpErrorHandler