--- a/framework/src/sbt-plugin/src/main/scala-sbt-1.0/play/sbt/run/PlaySource.scala
+++ b/framework/src/sbt-plugin/src/main/scala-sbt-1.0/play/sbt/run/PlaySource.scala
@@ -1,7 +1,6 @@
 /*
  * Copyright (C) 2009-2017 Lightbend Inc. <https://www.lightbend.com>
  */
-
 // This is a naive way to make sbt.internal.io.Source accessible. That is why we
 // are declaring the package here as sbt.internal.io. You can see the code for
 // sbt.internal.io.Source here: