Fixes #8089

This refactors the request handling code in the `AkkaHttpServer` and `NettyServer` slightly, so that the action is always executed in a future using the Play default execution context, then that future is flatMapped with a function that runs the body accumulator, and the resulting future is recovered using the error handler.

I added a test that tests with both the `Filter` and `EssentialFilter` case.