@schmitch 2017-12-12
does this has a performance impact?