@gmethvin 2017-12-13
Okay so I ran `Play-Microbenchmark/jmh:run -i 5 -wi 5 -f1 -t4 play.microbenchmark.it.HelloWorldBenchmark`

Before:

```
[info] Benchmark                       (endpoint)  (reqsPerConn)   Mode  Cnt      Score      Error  Units
[info] HelloWorldBenchmark.helloWorld   nt-11-pln              5  thrpt    5  24407.809 ± 3284.289  ops/s
[info] HelloWorldBenchmark.helloWorld   nt-11-enc              5  thrpt    5  24058.859 ± 9721.483  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-11-pln              5  thrpt    5  20829.463 ± 9841.523  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-11-enc              5  thrpt    5  14978.026 ± 2929.331  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-20-enc              5  thrpt    5   8906.489 ± 5659.373  ops/s
```

After:

```
[info] Benchmark                       (endpoint)  (reqsPerConn)   Mode  Cnt      Score      Error  Units
[info] HelloWorldBenchmark.helloWorld   nt-11-pln              5  thrpt    5  24209.883 ± 4650.606  ops/s
[info] HelloWorldBenchmark.helloWorld   nt-11-enc              5  thrpt    5  22155.185 ± 9727.310  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-11-pln              5  thrpt    5  19364.448 ± 6954.653  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-11-enc              5  thrpt    5  14367.340 ± 7419.241  ops/s
[info] HelloWorldBenchmark.helloWorld   ak-20-enc              5  thrpt    5   8522.170 ± 2350.906  ops/s
```

It seems like it may have slowed things down very slightly but it's hard to tell.