@schmitch 2017-12-13
well I would say this is fine. it's still in the derivation.