@gmethvin 2017-12-13
backport to 2.6.x: 3d784e9